Distribute x509_parser contents over remaining modules
